backend/requirements.txt
fastapi
uvicorn
pymongo
requests
pandas
python-dotenv